+++ src/random_generator/random_generator.hpp
@@ -10,7 +10,7 @@
 
 namespace random_generator
 {
-  typedef std::uniform_real_distribution<double> real_dist;
+  typedef std::uniform_real_distribution<float> real_dist;
   typedef std::uniform_int_distribution<> int_dist;
   typedef std::normal_distribution<> normal_dist;
